# swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "varsion": "0.0.1",
    "title": "DAS SDL API server",
    "description": "DAS SDL 서버 API swagger 문서",
    "version": "1.0.0"
  },
  "host": "localhost:3200",
  "basePath": "/das",
...
  "paths": {
...
    "/comm/notiInfo": {
      "get": {
        "tags": [
          "comm"
        ],
        "summary": "공지사항 조회",
        "description": "사용자가 공지사항을 조회한다",
        "parameters": [
          {
            "name": "page_no",
            "description": "페이지 번호",
            "in": "query",
            "type": "number",
            "example" : 1,
            "required" : true
          },
          {
            "name": "notice_title",
            "description": "공지사항 제목",
            "in": "query",
            "type": "string",
            "example" : "공지사항"
          },
          {
            "name": "user_name",
            "description": "등록자 이름",
            "in": "query",
            "type": "string",
            "example" : "관리자"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "공지사항 조회 성공"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "noti_id" : {
                        "type" : "number",
                        "example" : 15
                      },
                      "notice_title" : {
                        "type" : "string",
                        "example" : "공지사항제목"
                      },
                      "rgst_user_id" : {
                        "type" : "number",
                        "example" : 6
                      },
                      "user_name" : {
                        "type" : "string",
                        "example" : "홍길동"
                      },
                      "rgst_dtm" : {
                        "type" : "string",
                        "example" : "2024-08-29 16:58:09"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/comm/addNoti": {
      "post": {
        "tags": [
          "comm"
        ],
        "summary": "공지사항 등록",
        "description": "사용자가 공지사항을 등록한다",
        "parameters" : [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "notice_title": { 
                  "type": "string",
                  "example" : "공지사항_1", 
                  "required" : true
                },
                "notice_description" : {
                  "type": "string",
                  "example" : "공지사항 내용 어쩌구저쩌구", 
                  "required" : true
                },
                "rgst_user_id" : {
                  "type" : "number",
                  "example" : 1,
                  "required" : true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "공지사항 등록 성공"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/comm/detail": {
      "get": {
        "tags": [
          "comm"
        ],
        "summary": "공지사항 상세 조회",
        "description": "사용자가 공지사항의 상세화면을 호출한다",
        "parameters" : [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "notice_title": { 
                  "type": "string",
                  "example" : "공지사항_1", 
                  "required" : true
                },
                "notice_description" : {
                  "type": "string",
                  "example" : "공지사항 내용 어쩌구저쩌구", 
                  "required" : true
                },
                "rgst_user_id" : {
                  "type" : "number",
                  "example" : 1,
                  "required" : true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "object",
              "properties": {
                "resultCode": {
                  "type": "number",
                  "example": 200
                },
                "resultMsg": {
                  "type": "string",
                  "example": "공지사항 상세 조회 성공"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "userId" : {
                        "type" : "number",
                        "example" : 1
                      },
                      "loginId" : {
                        "type" : "string",
                        "example" : "altm56@gmail.com"
                      },
                      "isFirst" : {
                        "type" : "number",
                        "example" : 0
                      },
                      "userName" : {
                        "type" : "string",
                        "example" : "홍길동"
                      },
                      "userLevel" : {
                        "type" : "number",
                        "example" : 3
                      },
                      "userStatus" : {
                        "type" : "string",
                        "example" : "W"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    }
  }
}